changed command name, new Mobile
